Moved responses to separate file in order to use it in tests
