+++ features/responses/mentioning.js
@@ -1,0 +1,9 @@
+const responsesOnMentioning = [
+  "I'm listening",
+  'How can I help?',
+  'Yeap? :smile:',
+  'Mm? :smile:',
+  "I'm here :smile:",
+];
+
+module.exports = responsesOnMentioning;
